Final Changes
I redesigned most of the sprites to make better use of the grid space. My earlier sprites were too thin and didn't fill up the space properly. I was unable to complete the 100% task due to it being too challenging, but I made halfway progress and decided to include it anyway. My audio clips could be better, but they get the job done. I will likely change the audio clips in the future to be a little more 'food' themed. Overall, quite happy with the final result.