Adding basic features to the title screen
I've decided to include the old nurse sprites as well as the new ones in case the new ones are a little too abstract to get the point across. I used the square ones in the end because the older versions were too thin and causing problems with visual fidelity that I just couldn't solve no matter what. The text seems quite blurry on the game view and I'm not sure how to fix this.  So far it has the title, subtitle, pacman student, and nurses.