Update
I think I have basically finished the 50P portion of the assignment. I noticed that the "catchup" part of this assignment actually makes me regress a few elements, such as the cycling ghost animations and the pacman character running around the rectangle. It's likely for the later sections, but I've made those changes for now. I have to really get on some other assignments for other subjects, so there is a likelihood I will sit on this 50P pass unless I am fortunate enough to have extra time.